requirements.txt
Kodistubs==19.0.0
websocket-client==1.3.1
